fix(audio): match mobile app's preprocessing precision and bandpass filter
- Keep mel features and scaler stats in float64 through normalization,
  rounding to float32 once at the final TFLite input cast (already done
  in main.py/offline_score.py) instead of early. The app normalizes in
  double precision and only rounds once at the TFLite call; rounding to
  float32 before scaling (the old behavior) diverged from that ordering.
- Replace the two-stage highpass+lowpass filtfilt bandpass with a single
  combined Butterworth bandpass applied via one-pass lfilter, matching
  AudioDsp.applyBandpassFilter in the mobile app (scipy.signal.butter
  with btype="band" reproduces the app's hardcoded coefficients exactly).
- Default use_filter to False in main.py, offline_score.py, and
  FeatureExtractor.extract_features — the app ships with the filter off
  by default (opt-in via app settings) and most users never enable it.

Verified against a standalone Dart+fftea port of the app's mel-spectrogram
pipeline: the full feature-extraction output now matches to float32
rounding-noise level (~1 ULP on a small fraction of values).

+++ app/audio/scaler.py
@@ -64,13 +64,13 @@ class Scaler:
                 if mean is None:
                     return self._fail(f"{path}: JSON missing 'mean'/'mean_'")
 
-                mean = np.asarray(mean, dtype=np.float32)
+                mean = np.asarray(mean, dtype=np.float64)
 
                 # Prefer explicit var; otherwise compute from scale
                 if var is not None:
-                    var = np.asarray(var, dtype=np.float32)
+                    var = np.asarray(var, dtype=np.float64)
                 elif scale is not None:
-                    scale = np.asarray(scale, dtype=np.float32)
+                    scale = np.asarray(scale, dtype=np.float64)
                     var = scale * scale
                 else:
                     return self._fail(f"{path}: JSON missing 'var'/'scale'")
@@ -98,8 +98,8 @@ class Scaler:
                     f"{path}: .npz missing 'var'/'var_' (keys: {list(data.keys())})"
                 )
 
-            self.mean = np.asarray(mean, dtype=np.float32)
-            self.var = np.asarray(var, dtype=np.float32)
+            self.mean = np.asarray(mean, dtype=np.float64)
+            self.var = np.asarray(var, dtype=np.float64)
             return self.mean, self.var
         except Exception as e:
             return self._fail(f"{path}: {type(e).__name__}: {e}")
@@ -124,8 +124,8 @@ class Scaler:
         if mean is None or var is None:
             raise ValueError("Scaler not loaded: mean/var are None")
 
-        mean = np.asarray(mean, dtype=np.float32)
-        var = np.asarray(var, dtype=np.float32)
+        mean = np.asarray(mean, dtype=np.float64)
+        var = np.asarray(var, dtype=np.float64)
 
         scale = np.sqrt(var)
         # Avoid divide by zero
